Add shader variations, some refactorings
Activate transparency on the ShaderMaterial.

Create a function for each variation of the code to generate the wave
shader.

Add a shader variation which contains a sort of shadow in the bg color
as that area approaches the foreground one.

Add another shader variation starting from this last one to give the
illusion of depth by truncating the background color near the start and
the end of the material on the x axis.